NODE-665: Separate contexts introduced
